Leave only one benchmarks project
