+++ XReports.sln
@@ -16,13 +16,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Core.Tests", "test
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "benchmarks", "benchmarks", "{6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks.OldVersion", "benchmarks\XReports.Benchmarks.OldVersion\XReports.Benchmarks.OldVersion.csproj", "{D926E405-5165-4BFD-ACFB-B5F7B7CA707D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks.NewVersion", "benchmarks\XReports.Benchmarks.NewVersion\XReports.Benchmarks.NewVersion.csproj", "{3249B438-A9FA-4A15-8A52-4695B19ECF70}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks.Core", "benchmarks\XReports.Benchmarks.Core\XReports.Benchmarks.Core.csproj", "{0B54ECED-076D-4A2A-BC10-93FD160A105E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks.ConsoleApp", "benchmarks\XReports.Benchmarks.ConsoleApp\XReports.Benchmarks.ConsoleApp.csproj", "{43F53236-7373-4BE4-A846-3C0872C97C05}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Benchmarks", "benchmarks\XReports.Benchmarks\XReports.Benchmarks.csproj", "{3249B438-A9FA-4A15-8A52-4695B19ECF70}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XReports.Tests.Analyzers", "tests\XReports.Tests.Analyzers\XReports.Tests.Analyzers.csproj", "{D4D39377-AD89-4265-B9C2-8F105B5D7081}"
 EndProject
@@ -134,22 +128,10 @@ Global
 		{E3DDD5DB-6CD7-46AD-9DCA-F6D1D36E7714}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E3DDD5DB-6CD7-46AD-9DCA-F6D1D36E7714}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E3DDD5DB-6CD7-46AD-9DCA-F6D1D36E7714}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D926E405-5165-4BFD-ACFB-B5F7B7CA707D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D926E405-5165-4BFD-ACFB-B5F7B7CA707D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D926E405-5165-4BFD-ACFB-B5F7B7CA707D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D926E405-5165-4BFD-ACFB-B5F7B7CA707D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3249B438-A9FA-4A15-8A52-4695B19ECF70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3249B438-A9FA-4A15-8A52-4695B19ECF70}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3249B438-A9FA-4A15-8A52-4695B19ECF70}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3249B438-A9FA-4A15-8A52-4695B19ECF70}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0B54ECED-076D-4A2A-BC10-93FD160A105E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0B54ECED-076D-4A2A-BC10-93FD160A105E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0B54ECED-076D-4A2A-BC10-93FD160A105E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0B54ECED-076D-4A2A-BC10-93FD160A105E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{43F53236-7373-4BE4-A846-3C0872C97C05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{43F53236-7373-4BE4-A846-3C0872C97C05}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{43F53236-7373-4BE4-A846-3C0872C97C05}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{43F53236-7373-4BE4-A846-3C0872C97C05}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -271,10 +253,7 @@ Global
 		{45EFC336-0978-4C5C-91B6-22541C463736} = {F9E94722-061B-4A10-896F-880442F76D1F}
 		{10C08E20-75E1-41E5-B5EC-574F469770F7} = {0702E6BC-8F03-4672-972A-E16CBBD3424E}
 		{E3DDD5DB-6CD7-46AD-9DCA-F6D1D36E7714} = {F9E94722-061B-4A10-896F-880442F76D1F}
-		{D926E405-5165-4BFD-ACFB-B5F7B7CA707D} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
 		{3249B438-A9FA-4A15-8A52-4695B19ECF70} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
-		{0B54ECED-076D-4A2A-BC10-93FD160A105E} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
-		{43F53236-7373-4BE4-A846-3C0872C97C05} = {6E0C7C3A-81C6-453E-8FF9-1C3D96B97B88}
 		{D4D39377-AD89-4265-B9C2-8F105B5D7081} = {F9E94722-061B-4A10-896F-880442F76D1F}
 		{8DC2DB68-61B0-4ACA-BC5A-8DFFC560B1BB} = {F9E94722-061B-4A10-896F-880442F76D1F}
 		{FA27AF0A-5B27-4552-8946-09B5007A417A} = {1AE6B679-F900-4535-AAE4-B86D7B9652E3}
