--- benchmarks/XReports.Benchmarks.ConsoleApp/XReports.Benchmarks.ConsoleApp.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\XReports.Benchmarks.NewVersion\XReports.Benchmarks.NewVersion.csproj" />
-    <ProjectReference Include="..\..\src\XReports\XReports.csproj" />
-  </ItemGroup>
-
-</Project>
