--- benchmarks/XReports.Benchmarks.Core/Models/Address.cs
@@ -1,10 +1,0 @@
-namespace XReports.Benchmarks.Core.Models;
-
-public class Address
-{
-    public string Country { get; set; }
-    public string City { get; set; }
-    public string ZipCode { get; set; }
-    public string StreetAddress1 { get; set; }
-    public string StreetAddress2 { get; set; }
-}
