--- benchmarks/XReports.Benchmarks.Core/Models/Vehicle.cs
@@ -1,10 +1,0 @@
-namespace XReports.Benchmarks.Core.Models;
-
-public class Vehicle
-{
-    public string Manufacturer { get; set; }
-    public string Model { get; set; }
-    public string Vin { get; set; }
-    public string FuelType { get; set; }
-    public string Type { get; set; }
-}
