--- benchmarks/XReports.Benchmarks.Core/ReportStructure/Enums/Alignment.cs
@@ -1,8 +1,0 @@
-namespace XReports.Benchmarks.Core.ReportStructure.Enums;
-
-public enum Alignment
-{
-    Left,
-    Right,
-    Center,
-}
