--- benchmarks/XReports.Benchmarks.Core/ReportStructure/Models/Properties/BoldProperty.cs
@@ -1,14 +1,0 @@
-namespace XReports.Benchmarks.Core.ReportStructure.Models.Properties;
-
-public class BoldProperty : ReportCellsSourceProperty
-{
-    public override bool Equals(object obj)
-    {
-        return obj is BoldProperty;
-    }
-
-    public override int GetHashCode()
-    {
-        return this.GetType().GetHashCode();
-    }
-}
