--- benchmarks/XReports.Benchmarks.Core/ReportStructure/Models/Properties/ReportCellsSourceProperty.cs
@@ -1,7 +1,0 @@
-namespace XReports.Benchmarks.Core.ReportStructure.Models.Properties;
-
-public abstract class ReportCellsSourceProperty
-{
-    public abstract override bool Equals(object obj);
-    public abstract override int GetHashCode();
-}
