--- benchmarks/XReports.Benchmarks.Core/ReportStructure/Models/Properties/SameColumnFormatProperty.cs
@@ -1,14 +1,0 @@
-namespace XReports.Benchmarks.Core.ReportStructure.Models.Properties;
-
-public class SameColumnFormatProperty : ReportCellsSourceProperty
-{
-    public override bool Equals(object obj)
-    {
-        return obj is SameColumnFormatProperty;
-    }
-
-    public override int GetHashCode()
-    {
-        return this.GetType().GetHashCode();
-    }
-}
