--- benchmarks/XReports.Benchmarks.NewVersion/XReports.Benchmarks.NewVersion.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>disable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\XReports\XReports.csproj" />
-    <ProjectReference Include="..\XReports.Benchmarks.Core\XReports.Benchmarks.Core.csproj" />
-  </ItemGroup>
-
-</Project>
