--- benchmarks/XReports.Benchmarks.NewVersion/XReportsProperties/CustomFormatProperty.cs
@@ -1,7 +1,0 @@
-using XReports.Table;
-
-namespace XReports.Benchmarks.NewVersion.XReportsProperties;
-
-public class CustomFormatProperty : ReportCellProperty
-{
-}
