--- benchmarks/XReports.Benchmarks.OldVersion/Program.cs
@@ -1,12 +1,0 @@
-using BenchmarkDotNet.Configs;
-using BenchmarkDotNet.Engines;
-using BenchmarkDotNet.Jobs;
-using BenchmarkDotNet.Running;
-using XReports.Benchmarks.OldVersion;
-
-Job job = Job.Default
-    .WithStrategy(RunStrategy.Monitoring)
-    .WithIterationCount(1)
-    .AsDefault();
-IConfig config = DefaultConfig.Instance.AddJob(job);
-BenchmarkRunner.Run<ReportServiceBenchmarks>(config, args);
