--- benchmarks/XReports.Benchmarks.OldVersion/ReportServiceBenchmarks.cs
@@ -1,16 +1,0 @@
-using XReports.Benchmarks.Core;
-
-namespace XReports.Benchmarks.OldVersion;
-
-public class ReportServiceBenchmarks : BenchmarksBase
-{
-    protected override ReportService CreateReportService()
-    {
-        if (this.Data is null || this.Table is null)
-        {
-            throw new InvalidOperationException("Data or data reader is not initialized.");
-        }
-
-        return new ReportService(this.Data, this.Table);
-    }
-}
