--- benchmarks/XReports.Benchmarks.OldVersion/XReports.Benchmarks.OldVersion.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>disable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="XReports" Version="0.1.6" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\XReports.Benchmarks.Core\XReports.Benchmarks.Core.csproj" />
-  </ItemGroup>
-
-</Project>
