--- benchmarks/XReports.Benchmarks.OldVersion/XReportsProperties/CustomFormatProperty.cs
@@ -1,7 +1,0 @@
-using XReports.Models;
-
-namespace XReports.Benchmarks.OldVersion.XReportsProperties;
-
-public class CustomFormatProperty : ReportCellProperty
-{
-}
