--- benchmarks/XReports.Benchmarks.OldVersion/XReportsProperties/CustomFormatPropertyExcelHandler.cs
@@ -1,12 +1,0 @@
-using XReports.Models;
-using XReports.PropertyHandlers;
-
-namespace XReports.Benchmarks.OldVersion.XReportsProperties;
-
-public class CustomFormatPropertyExcelHandler : PropertyHandler<CustomFormatProperty, ExcelReportCell>
-{
-    protected override void HandleProperty(CustomFormatProperty property, ExcelReportCell cell)
-    {
-        cell.NumberFormat = "[=100]0;[<100]0.00";
-    }
-}
