--- benchmarks/XReports.Benchmarks.OldVersion/XReportsProperties/CustomFormatPropertyHtmlHandler.cs
@@ -1,16 +1,0 @@
-using System.Globalization;
-using XReports.Models;
-using XReports.PropertyHandlers;
-
-namespace XReports.Benchmarks.OldVersion.XReportsProperties;
-
-public class CustomFormatPropertyHtmlHandler : PropertyHandler<CustomFormatProperty, HtmlReportCell>
-{
-    protected override void HandleProperty(CustomFormatProperty property, HtmlReportCell cell)
-    {
-        decimal value = cell.GetValue<decimal>();
-        string format = value == 100m ? "F0" : "F2";
-
-        cell.Value = value.ToString(format, CultureInfo.CurrentCulture);
-    }
-}
