+++ benchmarks/XReports.Benchmarks/Models/Address.cs
@@ -1,0 +1,10 @@
+namespace XReports.Benchmarks.Models;
+
+public class Address
+{
+    public string Country { get; set; }
+    public string City { get; set; }
+    public string ZipCode { get; set; }
+    public string StreetAddress1 { get; set; }
+    public string StreetAddress2 { get; set; }
+}
