+++ benchmarks/XReports.Benchmarks/Models/Vehicle.cs
@@ -1,0 +1,10 @@
+namespace XReports.Benchmarks.Models;
+
+public class Vehicle
+{
+    public string Manufacturer { get; set; }
+    public string Model { get; set; }
+    public string Vin { get; set; }
+    public string FuelType { get; set; }
+    public string Type { get; set; }
+}
