+++ benchmarks/XReports.Benchmarks/XReportsProperties/CustomFormatProperty.cs
@@ -1,0 +1,7 @@
+using XReports.Table;
+
+namespace XReports.Benchmarks.XReportsProperties;
+
+public class CustomFormatProperty : ReportCellProperty
+{
+}
