+++ benchmarks/XReports.Benchmarks/XReportsProperties/CustomFormatPropertyHtmlHandler.cs
@@ -1,0 +1,16 @@
+using System.Globalization;
+using XReports.Converter;
+using XReports.Html;
+
+namespace XReports.Benchmarks.XReportsProperties;
+
+public class CustomFormatPropertyHtmlHandler : PropertyHandler<CustomFormatProperty, HtmlReportCell>
+{
+    protected override void HandleProperty(CustomFormatProperty property, HtmlReportCell cell)
+    {
+        decimal value = cell.GetValue<decimal>();
+        string format = value == 100m ? "F0" : "F2";
+
+        cell.SetValue(value.ToString(format, CultureInfo.CurrentCulture));
+    }
+}
